OLB-33 refactor: 지난달 보다 계산 중 past_date 구하는 방식 수정
