+++ app/module/chart/constant.py
@@ -6,6 +6,8 @@ RICH_PICK_SECOND = 60 * 60
 REDIS_RICH_PORTFOLIO_KEY = "richportfolio"
 REDIS_RICH_PORTFOLIO_SECOND = 60 * 60 * 24
 
+REDIS_STOCK_PAST_DATE_KEY = 'stock_past_date'
+
 DEFAULT_TIP = "자신이 사랑하는 일을 찾아 투자하세요."
 PAST_MONTH_DAY = 31
 FULL_PERCENTAGE_RATE = 100.0
